export tab details, clean up code
